Room occupancy planning service
* minor changes in API model
* created configuration and strategy for calculating occupancy and earnings
* added guest prices suppliers
* added custom assertions for RoomOccupancyPlan validations

+++ src/main/java/pl/s/h/interview/api/RoomAvailability.java
@@ -19,5 +19,5 @@ public record RoomAvailability(
 
         @NotNull
         @Min(0)
-        Integer availableRooms) {
+        long availableRooms) {
 }
